feat: rename php cs fixer config file
The `.dist`-less version is compatible with Laravel Shift.

BREAKING CHANGE: releasing as a new major version to communicate
this change involves the user making a change.

I know that it doesn't necessarily mean the user has to make a change,
but this is the closest thing to a major change that I can use to tell
semantic release to bump the major version of the package.